capemgr: Retry loading when failure to find firmware
When the rootfs device is on a device that's created via a dtbo
and the dtbo firmware is not including in the kernel image loading
failed.

The way we get around this is to retry the failed firmware loading
attempt until it is either successful, or no more capes are loading
(which means that the firmware file is never going to be found).

Signed-off-by: Pantelis Antoniou <panto@antoniou-consulting.com>